Fixing keyboard behaviour on AppState changes
Fixes - https://github.com/GetStream/stream-chat-react-native/issues/198

AppState detection logic is faulty. So far we had clear definition for change from background to foreground (in which case we setup the keyboard the listeners.), but we don't have clear definition of foreground to background (in which case we remove listeners), which resulted in keyboard listeners getting removed un-necessarily. Fixing it here by adding explicit check for transition from foreground to background.