refactor: change the direction gallery thumbnails are rendered
This applies to grids of 2x2 images.
Currently, we render images by column. This means that the first two
pictures are in column one, the last two in column two.

To conform to what the other SDKs does, this PR flips in by changing
`invertedDirection` to `true` for our use of `buildThumbnailGallery`.
This will render the photos by row instead.